+++ .github/workflows/scripts/win-test-services.ps1
@@ -3,7 +3,8 @@ param (
     [string]$access_token = "testing123",
     [string]$realm = "test",
     [string]$memory = "512",
-    [string]$with_fluentd = "true"
+    [string]$with_fluentd = "true",
+    [string]$with_msi_uninstall_comments = ""
 )
 
 $ErrorActionPreference = 'Stop'
@@ -51,5 +52,16 @@ if ("$with_fluentd" -eq "true") {
         throw "fluentdwinsvc service is running."
     } else {
         write-host "fluentdwinsvc service is not running."
+    }
+}
+
+if ($with_msi_uninstall_comments -ne "") {
+    $uninstallProperties = Get-ChildItem -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall" |
+        ForEach-Object { Get-ItemProperty $_.PSPath } |
+        Where-Object { $_.DisplayName -eq "Splunk OpenTelemetry Collector" }
+    if ($with_msi_uninstall_comments -ne $uninstallProperties.Comments) {
+        throw "Uninstall Comments in registry are not properly set. Found: '$uninstallProperties.Comments', Expected '$with_msi_uninstall_comments'"
+    } else {
+        write-host "Uninstall Comments in registry are properly set."
     }
 }